+++ src/main/java/net/joseph/vaultfilters/ItemAttributes/custom/NumberPrefixAttribute.java
@@ -80,7 +80,7 @@ public class NumberPrefixAttribute implements ItemAttribute {
                 i = 100000;
             }
         }
-        return Integer.parseInt(tempnum);
+        return Double.parseDouble(tempnum);
 
     }
